+++ src/mod_rm.cpp
@@ -1,9 +1,33 @@
 #include "mod_rm.hpp"
-#include "tinyexpr.h"
 #include "string_util.hpp"
 
 namespace nask_utility {
     namespace ModRM {
+
+        size_t get_imm_size_evenif_bracket(const std::string& hex_string) {
+            if (starts_with(hex_string, "[") && ends_with(hex_string, "]")) {
+                std::string reg = hex_string;
+                reg = reg.erase(0, 1);
+                reg = reg.erase(reg.size() - 1);
+                return get_imm_size(reg);
+            } else {
+                return get_imm_size(hex_string);
+            }
+        }
+
+        size_t get_imm_size(const std::string& hex_string) {
+            if (is_hex_notation(hex_string)) {
+                // サイズが偶数じゃないような場合の調整 ex) 0xffc => 0x0ffc
+                const std::string supplied_hex = (hex_string.substr(2).size() % 2 == 0) ?
+                    hex_string.substr(2) : "0" + hex_string.substr(2);
+                const size_t s = supplied_hex.size() / 2;
+                log()->info("check imm size: {} => {}", hex_string, s);
+                return s;
+            } else {
+                return 0;
+            }
+        }
+
         const std::string get_MMMSSS_from_reg(const std::string& reg) {
             if (REGISTERS_MMM_MAP.count(reg) != 0) {
                 return REGISTERS_MMM_MAP.at(reg);
@@ -211,5 +235,7 @@ namespace nask_utility {
             std::smatch match;
             return regex_match(reg, match, rm000);
         }
+
+
     };
 }
