+++ src/nask_utility.cpp
@@ -1449,7 +1449,7 @@ namespace nask_utility {
                 // JMP rel32 | 0xE9 cd
 
                 // offset = dst - src - current_pos
-                if (is_between_bytesize(token.AsLong()) && imm8 == get_imm_size(token.AsString())) {
+                if (is_between_bytesize(token.AsLong()) && imm8 == ModRM::get_imm_size(token.AsString())) {
                     // 0xEB
                     const long jmp_offset = (token.AsLong() - dollar_position - binout_container.size()) - 2;
                     log()->debug("JMP: {}", token.AsString());
@@ -1460,11 +1460,11 @@ namespace nask_utility {
                     // 0xE9
                     log()->debug("JMP: ", token.AsString());
                     const long jmp_offset = (token.AsLong() - dollar_position - binout_container.size()) - 3;
-                    if (get_imm_size(token.AsString()) == imm16) {
+                    if (ModRM::get_imm_size(token.AsString()) == imm16) {
                         log()->debug("0xe9 with Word {}", jmp_offset);
                         binout_container.push_back(0xe9);
                         set_word_into_binout(jmp_offset, binout_container);
-                    } else if (get_imm_size(token.AsString()) == imm32) {
+                    } else if (ModRM::get_imm_size(token.AsString()) == imm32) {
                         log()->debug("0xe9 with Dword {}", jmp_offset);
                         binout_container.push_back(0xe9);
                         set_dword_into_binout(jmp_offset, binout_container);
@@ -1868,7 +1868,7 @@ namespace nask_utility {
                     log()->debug("{} + {}", dst_reg, src_imm);
                     std::tuple<std::string, std::string> tp_dst = ModRM::REGISTERS_RRR_MAP.at(dst_reg);
                     // Imm8 or Imm16
-                    const size_t imm_size = get_imm_size(src_token.AsString());
+                    const size_t imm_size = ModRM::get_imm_size(src_token.AsString());
                     log()->debug("imm size: {}", imm_size);
 
                     if (ModRM::is_accumulator(dst_reg)) {
