+++ src/string_util.cpp
@@ -15,7 +15,7 @@ namespace nask_utility {
 
     bool is_imm8(const std::string& token) {
         if (is_hex_notation(token)) {
-            bool t = (get_imm_size(token) == imm8) ? true : false;
+            bool t = (ModRM::get_imm_size(token) == imm8) ? true : false;
             log()->info("{} is imm8 ? : {}", token, t);
             return t;
         } else {
@@ -103,30 +103,6 @@ namespace nask_utility {
         return s.compare(0, 2, "0x") == 0
             && s.size() > 2
             && s.find_first_not_of("0123456789abcdefABCDEF", 2) == std::string::npos;
-    }
-
-    size_t get_imm_size_evenif_bracket(const std::string& hex_string) {
-        if (starts_with(hex_string, "[") && ends_with(hex_string, "]")) {
-            std::string reg = hex_string;
-            reg = reg.erase(0, 1);
-            reg = reg.erase(reg.size() - 1);
-            return get_imm_size(reg);
-        } else {
-            return get_imm_size(hex_string);
-        }
-    }
-
-    size_t get_imm_size(const std::string& hex_string) {
-        if (is_hex_notation(hex_string)) {
-            // サイズが偶数じゃないような場合の調整 ex) 0xffc => 0x0ffc
-            const std::string supplied_hex = (hex_string.substr(2).size() % 2 == 0) ?
-                hex_string.substr(2) : "0" + hex_string.substr(2);
-            const size_t s = supplied_hex.size() / 2;
-            log()->info("check imm size: {} => {}", hex_string, s);
-            return s;
-        } else {
-            return 0;
-        }
     }
 
     std::ifstream::pos_type filesize(const char* filename) {
