+++ src/string_util.hpp
@@ -34,9 +34,6 @@ namespace nask_utility {
     bool is_contains_math_op(const std::string& subject);
     std::string replace_hex2dec(const std::string& subject);
     std::string expr_math_op(const std::string& subject);
-
-    size_t get_imm_size(const std::string& hex_string);
-    size_t get_imm_size_evenif_bracket(const std::string& hex_string);
 };
 
 #endif /* STRING_UTIL_HPP_ */
